Fix for two finger presses not working on 16k.
Before, if a key was already being pressed, it would not check for increased pressure to go from single to double press.
Now if the key is currently single pressed, it checks if the threshold for double press has been passed.
The delta threshold for 16k has also been increased, as the previous one was too low and was giving double press even for only one finger.